the most intuitive solution to the problem
You are in-charge of the cake for your niece's birthday and have decided the cake will have one candle for each year of her total age. When she blows out the candles, she’ll only be able to blow out the tallest ones.

For example, if your niece is turning  4 years old, and the cake will have 4 candles of height 3, 1, 2, 3, she will be able to blow out  candles successfully, since the tallest candle is of height 3 and there are 2 such candles.

The input for the function contains:
int n - number of candles
int[] ar - the heights of the candles
The output for the function contains:
Return the integer of the candles blown out.